Move DTOs to Greenfield namespace
