+++ Plugin/BTCPayServer.Plugins.Cashu/Controllers/Greenfield/GreenfieldCashuFailedTransactionsController.cs
@@ -4,8 +4,8 @@ using System.Threading.Tasks;
 using BTCPayServer.Abstractions.Constants;
 using BTCPayServer.Abstractions.Extensions;
 using BTCPayServer.Client;
+using BTCPayServer.Plugins.Cashu.Controllers.Greenfield.DTOs;
 using BTCPayServer.Plugins.Cashu.Data;
-using BTCPayServer.Plugins.Cashu.Models;
 using BTCPayServer.Plugins.Cashu.Data.Models;
 using BTCPayServer.Plugins.Cashu.Services;
 using BTCPayServer.Services.Invoices;
