+++ Plugin/BTCPayServer.Plugins.Cashu/Controllers/Greenfield/GreenfieldCashuLightningSecretController.cs
@@ -3,8 +3,8 @@ using System.Threading.Tasks;
 using BTCPayServer.Abstractions.Constants;
 using BTCPayServer.Abstractions.Extensions;
 using BTCPayServer.Client;
+using BTCPayServer.Plugins.Cashu.Controllers.Greenfield.DTOs;
 using BTCPayServer.Plugins.Cashu.Data;
-using BTCPayServer.Plugins.Cashu.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
